split tests one file per command
