+++ tests/test_commands/test_version.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from .helpers import run
+
+
+def test_version():
+    actual = run(['version'])
+    assert actual.count('.') == 2
